requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.1
onnxruntime-gpu==1.18.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
Pygments==2.18.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.0
six==1.16.0
sympy==1.13.2
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
